src/requirement.txt
Flask==3.0.0
mysql-connector-python==8.3.0
pandas==2.2.0
numpy==1.26.4
joblib==1.3.2
reportlab==4.0.7
scikit-learn==1.3.2
Werkzeug==3.0.1
python-dotenv==1.0.1
